requirements.txt
aerich==0.6.3
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.5.0
APScheduler==3.9.1
arrow==1.2.2
asgiref==3.5.1
asn1crypto==0.24.0
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.11.1
bidict==0.22.0
bilireq==0.1.2
binaryornot==0.4.4
blinker==1.4
bs4==0.0.1
cached-property==1.5.2
certifi==2018.8.24
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
cloudscraper==1.2.48
colorama==0.4.4
configobj==5.0.6
cookiecutter==1.7.3
cryptography==2.6.1
dictdiffer==0.9.0
distlib==0.3.4
docopt==0.6.2
expiringdict==1.2.1
fastapi==0.73.0
feedparser==6.0.8
filelock==3.6.0
frozenlist==1.3.0
greenlet==1.1.2
h11==0.12.0
haruka-bot==1.3.6.post1
httpcore==0.14.7
httplib2==0.11.3
httptools==0.4.0
httpx==0.22.0
httpx-socks==0.7.3
idna==3.3
importlib-metadata==4.11.3
iso8601==0.1.16
Jinja2==3.1.2
jinja2-time==0.2.0
jsonpatch==1.21
jsonpointer==1.10
jsonschema==2.6.0
loguru==0.6.0
Markdown==3.3.7
MarkupSafe==2.1.1
multidict==6.0.2
nb-cli==0.6.6
nonebot-adapter-cqhttp==2.0.0b1
nonebot-adapter-onebot==2.0.0b1
nonebot-plugin-apscheduler==0.1.2
nonebot-plugin-covid19-news==0.2.8.3
nonebot_plugin_heweather==0.5.6
nonebot-plugin-htmlrender==0.0.4.6
nonebot-plugin-petpet==0.2.17
nonebot-plugin-setu==1.1.1
nonebot-plugin-status==0.3.0
ujson==5.3.0
msgpack==1.0.3
nonebot2==2.0.0b2
numpy==1.22.3
oauthlib==2.1.0
packaging==21.3
Pillow==8.4.0
pipenv==2022.5.2
PixivPy==3.7.0
platformdirs==2.5.2
playwright==1.21.0
poyo==0.5.0
prompt-toolkit==3.0.29
psutil==5.9.0
pyasn1==0.4.8
pydantic==1.9.0
pyee==8.1.0
pyfiglet==0.8.post1
Pygments==2.12.0
pygtrie==2.4.2
PyJWT==2.3.0
pymdown-extensions==9.4
pyparsing==3.0.8
pypika-tortoise==0.1.4
PySimpleSOAP==1.16.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.2
python-markdown-math==0.8
python-slugify==6.1.2
python-socketio==5.6.0
python-socks==2.0.3
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
pyzbar==0.1.9
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.8
scipy==1.8.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.17.1
text-unidecode==1.3
texttable==1.6.3
tinydb==4.7.0
toml==0.10.2
tomlkit==0.9.2
tortoise-orm==0.18.1
tqdm==4.64.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.24.1
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
wcwidth==0.2.5
websocket-client==0.57.0
websockets==10.1
yarl==1.7.2
zipp==3.8.0
